Separated main.dart and Homescreen.dart
